Added Gimp
First one that requires a wildcard for the bundleID as the version family is embedded in the bundled ID.

Application was verified via OSQuery on install and grabbing a v2 off their release catelog.

Might have to revisit this as there may be other bundleID's but until I hear reports otherwise leaving this rule set alone.